0037 - application.yml 에 각종 설정들 모아두기

+++ src/main/java/com/back/boundedContext/market/domain/OrderItem.java
@@ -26,7 +26,7 @@ public class OrderItem extends BaseIdAndTime {
 
   private int salePrice;
 
-  private int payoutRate = MarketPolicy.PRODUCT_PAYOUT_RATE;
+  private double payoutRate = MarketPolicy.PRODUCT_PAYOUT_RATE;
 
   public OrderItem(Order order, Product product, String productName, int price, int salePrice) {
     this.order = order;
